src/third_party/blink/web_tests/media/controls/overflow-menu-focus.html

<!DOCTYPE html>
<html>
<title>Overflow menu give focus back to last focused element when gaining focus</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/run-after-layout-and-paint.js"></script>
<script src="../media-controls.js"></script>
<video controls></video>
<script>
async_test(t => {

  const video  = document.querySelector('video');
  video.src = '../content/test.ogv';
  const menu = overflowMenu(video);
  enableTestMode(video);

  video.onloadedmetadata = t.step_func(() => {
    singleTapOnControl(overflowButton(video), t.step_func(() => {
      assert_true(isControlVisible(menu), 'Overflow menu should be visible');

      let lastElement = getFocusedElement(video);

      menu.focus();
      setTimeout(t.step_func_done(() => {
        assert_equals(getFocusedElement(video), lastElement, 'overflow menu should give focus back');
      }));
    }));
  });
});
</script>
</html>
